New x and o file
New ideas unable to add to previous(existing file)
Existing file needed to copy out some functions hence its persistence.
Created the umpire class with player 1 and player 2 as class variables that come as prompts.
Making the players class variables is so that only the umpire would need to be instantiated for now.
Possible change to this if bugs arise.