Changed the Mastermind class to module
The CodeBreaker could not inherit the play board from the mastermind module and the color board at the same time.
The colour board had a bug of reseting itself every time a round is played hence the move the the mastermind module.
The fix sets just one color board and play board for the game.