Added new methods to the auto code breaker class
Human feedback method allows the human to examine a particular location on the board and tell the computer how many red and white pegs it gets.
This is so that the human can be inclusive in the game as the umpire rather than automating the checker.
Work on feedback currently takes the human feedback and for now only has functionality if all colors selected match those of the human.