Added changes and bug fixes
Moved all the modules to a new file because requiring main in game start sequence file (to access the classes) and also requiring game start sequence in main, caused main to treat the modules as 'constants' and couldn't access them properly.
Creating the replies to wrong and correct guess methods is so as to DRY Out the code and be easily resused in the continue game method.