New bug fixes
Giving the hangmancontols class an instance is so that when saving the file, it would be able to access the name variable.
Previously this was not possble because probably the scope of the function that calls the save game method couldn' access the name variable if it was only defined the gamer class.